Revert "Move InMemoryTimerInternals to runners-core"
This reverts commit ec0bf7b4023ff75f4ec6723d2e77ed507eb57c51.
